refactor(nextly): publish the trigger authorizer the webhook drain proved out
R-2 adds a second scheduler trigger, and its auth question is the one the drain
already answers: a bearer secret for the scheduler, or a human with the right
permission over a POST from a same-origin context. The reasoning behind that --
why a GET may not reach the session path, why an API key is exempt from the
origin check, why the entropy floor is enforced at the boundary rather than only
in the env schema -- is worth more than the twenty lines that carry it, and a
second copy would be a second place for it to be wrong.

Moved rather than rewritten. What varies between triggers is the permission a
human needs, so that is the parameter; the webhook drain now names only that.

No behaviour change: the drain's 19 tests pass unmodified.
